Change type of Message.ident to int
Store the ident value as an integer instead of a byte string with
a length of one, which would inevitably cause trouble somewhere.

Also incidentally fixes an issue packing/unpacking messages in the
unlikely event that a message is greater than 2^31 bytes due to using
a signed integer type for the message size.